refactor: extract inline tests from oversized modules
Move large inline test blocks into sibling files so the source-size guardrails pass without changing coordination or instruction behavior.
